requirements.txt
pandas>=1.5.0
requests>=2.28.0
python-dotenv>=0.19.0
schedule>=1.2.0
pytz>=2022.1
numpy>=1.24.0